Move GenerationOptions to a better location and fix property name
