+++ exo/api/chatgpt_api.py
@@ -13,7 +13,8 @@ import signal
 from exo import DEBUG, VERSION
 from exo.helpers import PrefixDict, shutdown, get_exo_images_dir
 from exo.inference.tokenizers import resolve_tokenizer
-from exo.orchestration import Node, GenerationOptions
+from exo.orchestration import Node
+from exo.inference.generation_options import GenerationOptions
 from exo.models import build_base_shard, build_full_shard, model_cards, get_repo, get_supported_models, get_pretty_name
 from typing import Callable, Optional
 from PIL import Image
@@ -365,7 +366,7 @@ class ChatGPTAPI:
         shard,
         prompt,
         request_id=request_id,
-        generation_options=GenerationOptions(max_completion_tokens=chat_request.max_tokens)
+        generation_options=GenerationOptions(max_completion_tokens=chat_request.max_completion_tokens)
       ))), timeout=self.response_timeout)
 
       if DEBUG >= 2: print(f"[ChatGPTAPI] Waiting for response to finish. timeout={self.response_timeout}s")
