+++ exo/inference/generation_options.py
@@ -1,0 +1,8 @@
+from typing import Optional, List
+
+
+class GenerationOptions:
+  max_completion_tokens: Optional[int] = None
+
+  def __init__(self, max_completion_tokens: Optional[int] = None):
+    self.max_completion_tokens = max_completion_tokens
